refactor(velociraptor): move the message cap into the module that discloses it
The size ledger caps velociraptorImport.ts at 1,869 lines. Master's own work took
it to 1,868, so the 12 lines this branch adds put it over — and the gate's advice
is the right answer here rather than a workaround: fullMessage and MESSAGE_CAP
belong with the note that explains them, not beside the row mappers.

cappedMessage() now takes the raw message string instead of a Row, so the module
needs nothing from the importer. Behaviour is identical; velociraptorImport.ts
ends 12 lines BELOW where master left it.
